qt6: Cast qsizetype to int/unsigned in format strings
In Qt6, container size() returns qsizetype (long long) instead of
int. Cast to int or unsigned to match %d/%u format specifiers in
printf() and qDebug() calls.

Co-developed-by: Claude Opus 4.6 <noreply@anthropic.com>

+++ filemax.cpp
@@ -2845,8 +2845,8 @@ err_info *Filemax::show_file (FILE *f)
 
    fprintf (f, "\nfile: %s\n", _filename.toLatin1 ().constData());
    fprintf (f, "chunk0 start %x\n", _chunk0_start);
-   fprintf (f, "chunk count %d\n", _chunks.size ());
-   fprintf (f, "page count %d\n", _pages.size ());
+   fprintf (f, "chunk count %d\n", (int)_chunks.size ());
+   fprintf (f, "page count %d\n", (int)_pages.size ());
    fprintf (f, "signature %x\n", _signature);
 
    fprintf (f, "\nPage Summary:\n");
